chore(build): enable `-Wconversion` warning and remove some implicit type conversions

+++ libraries/Duet3D/General/StringRef.cpp
@@ -281,13 +281,13 @@ bool StringRef::Insert(size_t pos, const char* s) const noexcept
 int StringRef::Contains(const char* s) const noexcept
 {
 	const char* const r = strstr(p, s);
-	return (r == nullptr) ? -1 : r - p;
+	return (r == nullptr) ? -1 : static_cast<int>(r - p);
 }
 
 int StringRef::Contains(char c) const noexcept
 {
 	const char* const r = strchr(p, (int)c);
-	return (r == nullptr) ? -1 : r - p;
+	return (r == nullptr) ? -1 : static_cast<int>(r - p);
 }
 
 // Replace the first instance of oldVal by newVal, returning true if a replacement was done
